moved Pinpiont Blokcs out of modular and removed debuging code from other files
