+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/PinpointBlocks.java
@@ -20,14 +20,13 @@
  *   SOFTWARE.
  */
 
-package org.firstinspires.ftc.teamcode.modular;
+package org.firstinspires.ftc.teamcode;
 
 import org.firstinspires.ftc.robotcore.external.BlocksOpModeCompanion;
 import org.firstinspires.ftc.robotcore.external.ExportToBlocks;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.Pose2D;
-import org.firstinspires.ftc.teamcode.GoBildaPinpointDriver;
 
 import java.util.List;
